TechSolutions/requirements.txt
tkinter
mysql-connector-python
reportlab
python-dotenv